fix: change modifiedgame from c_int to c_uint to match C boolean type
Agent-Logs-Url: https://github.com/sunsided/room/sessions/8ad03347-8437-4911-a1dc-fafc22283ead

+++ room/src/doom/d_main.rs
@@ -1578,7 +1578,7 @@ pub extern "C" fn D_DoomMain() {
         }
 
         // Load PWAD files
-        modifiedgame = W_ParseCommandLine();
+        modifiedgame = W_ParseCommandLine() as c_uint;
 
         // Check for -playdemo / -timedemo
         let p = M_CheckParmWithArgs(b"-playdemo\0".as_ptr() as *const c_char, 1);
